Work for initial release (big change)
- Remove unneeded files (options/popup/styles/storage).
- Clean up build task (gulp) to remove SCSS.
- Remove promo resources (could be generated as needed).
- Modify local/manifest/package/readme to suite changes.
- Use yarn (thus yarn.lock).
- Update icon/images in sketch.
- Add usage screenshot.
- Add main functionality in background/contentscript files.

GitHub treats pages and loads as a single page application, so we need
to use the `webNavigation.onHistoryStateUpdated` to get the URL changes.
The problem is that it would fire twice per page change, so to counter
this we have to use a page tracker and to limit the button addition to
only one change.