Open TMWIC.com in a new window / tab
Currently the extension will navigate the current window away from
GitHub when users click on the "Tell Me When It Closes" button.

This change updates the behavior to open a new window/tab when a user
clicks on the "Tell Me When It Closes" button so that the context of the
GitHub issue is not lost.

This could be exposed as an option, although I'm generally in favor of
avoid configuration whenever we can.