Use a link to allow opening in new tab
Currently the extension uses a button to provide the "Tell Me When It
Closes" connection. While the button visually matches the GitHub UI,
using a button and JS `location.url = ...` to direct the user to
https://tellmewhenitcloses.com prevents the user from opening the page
in a new tab, should they want to.

This change updates to use a link instead of the current button.
Visually and functionally things are the same, but now users can opt to
cmd-click or shift-cmd-click to control how the page opens.